add profile modification,s3 storage minio

+++ src/module/cron/cron.service.ts
@@ -1,6 +1,7 @@
 import { PrismaService } from '@/src/core/prisma/prisma.service';
 
 import { MailService } from '../libs/mail/mail.service';
+import { StorageService } from '../libs/storage/storage.service';
 
 //import { StorageService } from '../libs/storage/storage.service';
 
@@ -12,7 +13,7 @@ export class CronService {
 	constructor(
 		private readonly prismaService: PrismaService,
 		private readonly mailService: MailService,
-		//private readonly storageService: StorageService,
+		private readonly storageService: StorageService,
 	) {}
 
 	@Cron('0 0 * * *')
@@ -31,9 +32,9 @@ export class CronService {
 		for (const user of deactivatedAccounts) {
 			await this.mailService.sendAccountDeletion(user.email);
 
-			/* 	if (user.avatar) {
+			if (user.avatar) {
 				await this.storageService.remove(user.avatar);
-			} */
+			}
 		}
 
 		await this.prismaService.user.deleteMany({
